requirements.txt
thulac
certifi==2019.3.9
chardet==3.0.4
Django==2.2
djangorestframework==3.9.2
djangorestframework-stubs==0.4.2
idna==2.8
jieba==0.39
numpy==1.16.2
PyMySQL==0.9.3
pytz==2019.1
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
sqlparse==0.3.0
urllib3==1.24.1
